benchmarks: replace causaltensor with the authors' R for SDID and MCNNM
Removes the causaltensor (Python) cross-checks entirely and makes the authors'
own R the sole reference for the canonical sdid_prop99 / mcnnm_prop99 cases.

- sdid_prop99: cross-validates mlsynth's SDID against a pinned live run of
  synth-inference/synthdid (synthdid_estimate). ATT -15.605 vs -15.604, matched
  to 1.6e-3; the package's DiD/SC estimates recorded for context.
- mcnnm_prop99: cross-validates MCNNM against susanathey/MCPanel (mcnnm_cv,
  defaults). ATT agrees to 0.15 packs, California counterfactual path to RMSE
  0.47; loose by construction and documented (the engines agree to ~3e-3 at a
  matched threshold; the gap is CV-penalty selection + the extrapolated block).

The former sdid_prop99_synthdid / mcnnm_prop99_mcpanel cases and reference
bundles are folded into the canonical names. causaltensor is dropped from the
registry, benchmarks README, the agents benchmarking guide, compare.py, and the
export_comparison package-key list (which now recognises synthdid / MCPanel).
Replication and estimator docs updated to cite the authors' R throughout; the
stale causaltensor comparison.csv exports are removed. Cases read the pinned
reference.json, so no R is needed to run them.
