+++ benchmarks/compare.py
@@ -6,7 +6,7 @@ from typing import Dict, Tuple
 
 class BenchmarkSkipped(Exception):
     """Raised by a case's ``run()`` when an optional dependency or reference
-    implementation is unavailable (e.g. ``causaltensor`` not installed, or the
+    implementation is unavailable (e.g. a reference package not installed, or the
     reference repo could not be cloned). The driver reports it as ``SKIP`` --
     distinct from a numerical ``FAIL`` -- so a missing optional toolchain never
     turns the suite red.
