+++ benchmarks/reference/mcnnm_prop99/reference.json
