+++ benchmarks/reference/mcnnm_prop99/reference.out
@@ -1,0 +1,2 @@
+MCPanel mcnnm ATT = -19.980463
+wrote benchmarks/reference/mcnnm_prop99/reference.json
