--- benchmarks/reference/mcnnm_prop99_mcpanel/reference.out
@@ -1,2 +1,0 @@
-MCPanel mcnnm ATT = -19.980463
-wrote benchmarks/reference/mcnnm_prop99_mcpanel/reference.json
