+++ benchmarks/reference/sdid_prop99/reference.out
@@ -1,4 +1,4 @@
 SDID ATT = -15.603829
 DID  ATT = -27.349111
 SC   ATT = -19.619665
-wrote benchmarks/reference/sdid_prop99_synthdid/reference.json 
+wrote benchmarks/reference/sdid_prop99/reference.json
